Merged PR 49: fix-docker-build
- Revert "Merged PR 47: Revert changes to Docker build process"
This reverts commit e01af4af178b2530dea324fe4ae74a64bf3f4fde.

- setup: fix docker build with dev profile
Fix the build path for 'docker_build_wrapper.sh dev' command.
$TOP_DIR refers to project absolute path on host only and must not be
used in the container.

- setup: refactor directory path management
The number of scripts, relying on install path, is increasing and they
are prone to be run on either a host or or a container. All the path are
now set and sourced from a single point for easier maintenance.