Merged PR 144: Add RTP sequence number analyzer implementation
This implementation comes with several Unit Tests and is also integrated with the Video, Audio and Metadata stream handlers.

The sequence number analyzer (`ebu_list::rtp::sequence_number_analyzer`) tracks the number of dropped packets in a sequence of RTP packets. It maintains a list (`std::vector<uint32_t>`) of the upper/lower bounds of the continuous sequences that were received. In the end, if there are any holes between those sequences then some packets were lost. Note that:
- It assumes the sequence has well defined upper and lower bounds for the sequence number.
- It tolerates packets that come out of order.

Currently not implemented but easy to add:
- List the sequence numbers of the packets that we're processed/lost.
- List the sequence numbers of the packets that we're repeated.

Currently not implemented and requires further discussion to so:
- Detecting when the sequence number is restarted and associate it to a new sequence. How can we do this? Maybe taking into account the RTP timestamp?

Related work items: #298