Merged PR 166: 329 - No PDF report button
There were 2 main issues regarding #329 bug report:
- Missing button to export a report as PDF.
- Exception being thrown in the `listwebserver` when exporting the report as JSON.

I wasn't able to reproduce the first one. Yet it seems to me that it was caused by the browser was using a cached version of the GUI, therefore using an URL that no longer exists.
This led me to the fix that I'm PRing: don't allow an invalid report URL to crash the `listwebserver`.

Related work items: #329