Merged PR 216: Update SDPoker version
listwebserver: Update SDPoker version

Fixed by ricardo@bisect.pt

This version of SDPoker should now validate the existence of a whitespace
character after the 'source-filter' attribute and before the ':'.

At the time of this commit there hasn't been a release of SDPoker where
this fix is included, so for now we're pointing to the specific commit
where it is available.

Related work items: #384