Merged PR 231: Improve audio RTP-to-pkt analysis
TSDF is updated every 200ms and RTP-to-pkt is measured for every
incoming packet. The single influx logger logs TSDF and RTP stat
(min, max. avg) every 200ms. If a capture is < 200ms, there is no
data logged at all.

The idea is to split the logger into rtp + tsdf so that they can be called
at their respective rate. The 2 loggers are still attached to the same
stream handler because RTP-to-pkt is used by TSDF calculus. There is
no more need to calculate stat for RTP but just keep the raw value of
RTP-to-pkt.

In this way, a very short capture (< 200ms) still contains RTP analysis.

Related work items: #341