Merged PR 250: Clarify video and audio results.
- doc: refactor video analysis
Merge the 2 doc files, since they both deal with timings and -21.
Homogenize the the vocabulary.
Add a link in the UI.
- gui: swap panels for video and audio panels
This new order is more appropriate: from generic info to media-specif
info.
- gui: relocate doc link for ancillary and video
- gui: display color depth
- gui: display analysis results in a dedicated pane
This allow to emphasize the measurement vs pass criteria.
- gui: use generic result pane for audio
- video: fix range for for delta between RTP TS and NxTframe
Align with JT-NM tested program [1] 4.4:
"The instantaneous value of theRTP timestamp of the stream
is not in the future, not more than 1 ms in the past"
[1] http://jt-nm.org/documents/JT-NM_Tested_Catalog_ST2110_Full-Online-2019-09-10.pdf
- gui: display errors in red
  * Pass complete stream info to Dash21Info.js to be able to parse any
analysis results.
  * Use the 'attention' key to select the red font color
  * Apply to video, audio and network analysis
  * Rename a couple of variables to camelCase
- gui: display TRoffset values in VRX and Tvd graphes
- video: clean RTP post analysis
  * cleanup names
  * add limit just like audio
  * display limits in the UI
- audio: align post analysis with video
  * cleanup
  * use units
  * update doc
- audio: replace sample_size with packet_size
- audio+video: display 'not compliant' in red in summary
- audio: rename all the pkt-to-rtp stuff
- video: clarify RTP analysis
- gui: remove graph: 'Delta from RTP TS to Packet TS'
We already have 'Delta 1st packet to RTP' in us, so keeping this one (in
ticks) confuses more than clarifies.
- gui: update translations

Related work items: #391, #427, #448, #449, #450