Merged PR 248: ST2110 Extractor - Refactoring and Data Layer
This PR moves a few generic components to the `apps_common` library (ex: serializers, `png_writer`, ...) and creates a layer between the `st2110_extractor` core logic and its access to data, for both reading and writing operations.
The previous implementation would use Mongo and Influx databases directly as storage. This PR adds the possibility to use the _filesystem_ as another option for storage.

Since the `st2110_extractor` also needs to read the metadata about the network capture and streams to work properly, the data needs to be fead through a pair of _JSON_ files. See #431 attachments for a concrete example.

Related work items: #429, #430, #431